feat: use local settings table, only persist to SHADA on exit
- Change to use a local table for settings and only put in an all-caps vim
global upon `VimLeavePre` so it gets persisted to SHADA before exiting. This
simplifies the code a bit in a number of areas.

- Fix: failing to restore settings from SHADA when the module is sourced/required
from init.lua.  This occured because SHADA is read in after user's init.lua.
So checking for the plugin's all-caps vim global upon requiring the module and
if not there creating an `VimEnter` autocmd to check for it again after SHADA
loads.